+++ src/contexts/Guilds/config/context.ts
@@ -1,6 +1,6 @@
 import React from 'react';
-import { GuildServiceProviderState } from './configProvider';
 import { Web3ReactContextInterface } from '@web3-react/core/dist/types';
+import { GuildServiceProviderState } from './provider';
 
 export interface ConfigContext extends GuildServiceProviderState {
   initialize: (context: Web3ReactContextInterface, address: string) => void;
